Added Tool / Job insertion. No validation on entered data so far. Need to check how it works on real device

+++ app/src/main/java/com/example/gyrodatasakhalin/job/JobItem.kt
@@ -9,7 +9,7 @@ data class JobItem(
     @SerializedName("Bullplug")
     val bullplug: String,
     @SerializedName("CirculationHours")
-    val circulationHours: Int,
+    val circulationHours: Float,
     @SerializedName("ClientName")
     val clientName: String,
     @SerializedName("Comment")
